feat(pvp): rodada de balanceamento das classes no torneio de 14 variações
O torneio (660 duelos, cada variação contra cada outra) passa a incluir as
DUAS 8ªs do BeastMaster que faltavam — Elemental e Evocação. Um personagem
aprende uma 8ª só, então elas são três BMs diferentes e têm de disputar
separadas. Foi o que revelou o desequilíbrio maior da lista abaixo.

Antes, três correções de INSTRUMENTO, sem as quais os números mentiam:

- o TK Trans lutava só no soco. As três skills de dano da árvore dele nunca
  entraram na simulação: medidas contra o Paladino, a Carga dá 26.230, o Golpe
  Mortal 23.424 e a Espada da Fênix 24.664, contra 805 do golpe normal. Era
  por isso que o placar dele não se movia com ataque nenhum — de 0 a 130
  pontos de multiplicador, sempre as mesmas 30 vitórias;
- a FM Cancelamento de arco lutava sem o Cancelamento e sem a Névoa. Trocar a
  arma não tira dela a árvore;
- o botão de dano da FM Cancelamento era invisível: a busca binária que resolve
  a ficha para bater a janela medida em jogo reabsorvia qualquer valor. Agora
  ela calibra no valor de referência e devolve o delta depois.

As mudanças de regra, todas varridas por VITÓRIAS no torneio, não por dano por
segundo (o dano por segundo médio esconde a esquiva e a parede da poção):

- BM Natureza, ponta de FORÇA: o eixo ganha a faixa de ACERTO que faltava
  (0 na Destreza pura, 140‰ na Força pura). Ele tinha a melhor ficha do elenco
  e era esquivado 33% das vezes contra 5% do irmão de Destreza — a Destreza
  tinha duas funções de graça, esquivar e furar esquiva. 29 → 68 vitórias;
- TK Trans: a Armadura Crítica passa a pagar ataque, que nunca pagou — até
  +250 de multiplicador e 30% de perfuração, os dois pela Força — e a defesa
  dela sobe de 100‰ para 600‰. Ele não perdia por bater pouco: morria em 126 s
  contra um Paladino que levaria 200 s para matar. 18 → 43;
- FM Cancelamento física: dano de duas armas 100 → 250 e a Névoa Venenosa vale
  +150% com a empunhadura de assinatura. A Névoa rendia 301 por uso contra 610
  do soco dela — ela gastava um terço dos turnos numa skill pior que bater.
  15 → 56;
- TK Espada Mágica: a lança 140% → 190%. Ele tirava 2.582/s contra os 3.603 da
  Black e levava 2.392 com uma poção que só levanta 2.000. 34 → 56, e o duelo
  dele passa de 20 s para 73 s;
- BM Elemental: lança 120% → 210% e cajado de duas mãos 140% → 250%. A ficha
  dele é a MESMA da Black, item por item; a diferença inteira é que ela tem o
  crítico de mago e ele não, por decisão de quem opera. Só que ele é justamente
  o BM que abriu mão do bando, e ficava sem o crítico E sem as criaturas.
  10 → 63;
- BM Evocação: o bando da Súcubo tirava 3.467/s SOZINHO, com o dono parado,
  contra uma poção de 2.000 — 72% do dano dele, e paralelo, porque não ocupa o
  turno de ninguém. O teto do bando é a poção. Súcubo 2.600 → 1.390 e Dragão
  Negro 1.320 → 1.104 (este para a 8ª não ficar valendo menos que a 7ª).
  117 → 59. As seis primeiras criaturas não mudam: quem não tem a 8ª nem
  alcança a maestria cheia em que a tabela vale.

Doze das catorze variações agora cabem entre 68 e 31 vitórias; antes o
intervalo ia de 117 a zero.

Também: teto de 100% na soma das perfurações, que hoje nunca se somam (são de
classes diferentes) mas acima de 100 dariam defesa NEGATIVA, multiplicando o
golpe em vez de deixá-lo passar.

+++ tmserver/internal/handler/simulacao_bm_test.go
@@ -125,7 +125,7 @@ func (sm *simulador) bmLutador(id int, urso bool, criatura int) *lutador {
 		// O bando bate em paralelo, na cadência dele, faça o dono o que fizer.
 		bando := 0
 		if agora >= proximoBando {
-			proximoBando = agora + evocacaoCadenciaMs
+			proximoBando = agora + int64(evocacaoCadenciaMs)
 			evo := int(effectiveSpecial(ld.e, 2))
 			n := summonCount(criatura+1, evo)
 			porCabeca := danoDaEvocacaoEmJogador(criatura, evo, int(effectiveAC(alvo)))
